requirements.txt
# ══ NexusSniff — Dependencias Python ══

# Framework UI
PyQt6>=6.6.0
pyqtgraph>=0.13.3

# Bases de datos
psycopg2-binary>=2.9.9
clickhouse-driver>=0.2.6
redis>=5.0.0

# Build del motor C++
pybind11>=2.12.0

# Utilidades
numpy>=1.26.0

# Exportación
openpyxl>=3.1.0
